refactor(workroom): give the outcome-anchor resolver its own module
The module-size ratchet failed: mcp-handlers.ts grew 847 -> 870. Re-baselining
would have been the wrong answer — that file is ALREADY baselined as oversized,
so adding to it moves the ratchet in the direction it exists to prevent.

Moved backlogItemIdFromOutcomeAnchor and its rationale into
lib/work-capsules/outcome-anchor.ts. That reclaims 21 of the 23 lines; the
irreducible remainder is one import and one field on the adoption input, so the
baseline moves 847 -> 849 with that justification.

Hand-edited the single baseline line rather than running --update, which
retightened NINE unrelated files that had shrunk since the baseline —
agentic-loop.ts 2806 -> 2704, TopologyGraph.tsx 973 -> 906, and seven more.
Carrying those would have put unrelated ratchet changes in a one-concern PR and
silently constrained other people's files against a snapshot of this branch.

Build gate re-run after the move: 20 test files / 133 tests passing across
lib/work-capsules; tsc --noEmit clean; pnpm --filter web build exit 0; Module
Size Guard green; all 7 pull-request policy guards green.

Note on the source profile: PR Health Logic and Janitor Tests fail locally and
fail identically on unmodified main — Windows host limits (symlink creation
needs admin, no executable bit). CI runs Linux and reported only Module Size
Guard, 1 of 56, which this fixes.
