+++ scripts/module-size-baseline.txt
@@ -61,7 +61,7 @@ apps/web/lib/tak/agent-routing.ts	1191
 apps/web/lib/tak/agentic-loop.test.ts	2501
 apps/web/lib/tak/agentic-loop.ts	2806
 apps/web/lib/tak/route-context-map.ts	1440
-apps/web/lib/work-capsules/mcp-handlers.ts	847
+apps/web/lib/work-capsules/mcp-handlers.ts	849
 apps/web/lib/work-capsules/work-capsule-store.test.ts	1191
 apps/web/lib/work-capsules/work-capsule-store.ts	1052
 apps/web/lib/workbooks/platform-tables.ts	938
